Fix compiler warnings (#3583)
Includes:
* type casting
* properly counting fields to initialise

+++ src/ride/gentle/crooked_house.c
@@ -162,7 +162,7 @@ static void sub_88ABA4(uint8 direction, uint8 x_offset, uint8 y_offset, uint32 s
 	RCT2_GLOBAL(0x0014281B8, uint32) = ride_type->vehicles[0].base_image_id;
 
 	if (ride->lifecycle_flags & RIDE_LIFECYCLE_ON_TRACK) {
-		if (ride->vehicles[0] != -1) {
+		if (ride->vehicles[0] != (uint16)-1) {
 			rct_sprite *sprite = &g_sprite_list[ride->vehicles[0]];
 			RCT2_GLOBAL(RCT2_ADDRESS_PAINT_SETUP_CURRENT_TYPE, uint8) = VIEWPORT_INTERACTION_ITEM_SPRITE;
 			RCT2_GLOBAL(0x9DE578, rct_sprite *) = sprite;
